[REFACTOR] Separate files into align, data, structure packages
