+++ pyprot/align/align.py
@@ -1,146 +1,12 @@
 from copy import deepcopy
 
-from pyprot.base import protein
-
-
-class Aligned:
-    """
-    Represents two aligned sequences with some metadata about the alignemnt.
-    """
-
-    def __init__(self, seqA, seqB, seqAStart, seqBStart, alignType, alignScore, scoreMatrix, isMultiple=False):
-        assert (len(seqA) == len(seqB))
-        # seqA is always a Sequence, seqB can be a list of ints (if MSA)
-        self.seqA = seqA if isinstance(seqA, Sequence) else seqB
-        self.seqB = seqB if isinstance(seqA, Sequence) else seqA
-        self.seqAStart = seqAStart if isinstance(seqA, Sequence) else seqBStart
-        self.seqBStart = seqBStart if isinstance(seqA, Sequence) else seqAStart
-        self.size = len(self.seqA)
-
-        self.alignType = alignType  # Options used for alignemnt
-        self.alignScore = round(alignScore, 2)  # Score of alignment
-
-        self.isMultiple = isMultiple  # Is this a M.S.A ?
-        self.pssmDescription = None
-        if self.isMultiple:
-            self.pssmDescription = scoreMatrix.getDescription()
-
-        self.condensed = False  # When true, aligned sequences are not displayed
-        self.chunkSize = 80  # Used for display
-        self.sizeIndic = 20  # Same
-
-        self.seqInter = []  # Sequences separators (.: )
-        self.identity = 0  # Identity (same AAs)
-        self.gaps = 0  # Gap count (either AA is a gap)
-        self.seqAGaps = 0
-        self.seqBGaps = 0
-        self.similarity = 0  # Similarity (positive score between AAs)
-
-        if self.isMultiple:
-            for aa in self.seqA:
-                if aa.isGap():
-                    self.gaps += 1
-                    self.seqAGaps += 1
-        else:
-            for aa1, aa2 in zip(self.seqA, self.seqB):
-                if aa1.isGap() or aa2.isGap():
-                    if aa1.isGap():
-                        self.seqAGaps += 1
-                    else:
-                        self.seqBGaps += 1
-                    self.seqInter.append(" ")
-                    self.gaps += 1
-                elif aa1 == aa2:
-                    self.seqInter.append(":")
-                    self.identity += 1
-                elif scoreMatrix.getScore(aa1, aa2) >= 0:
-                    self.seqInter.append(".")
-                    self.similarity += 1
-                else:
-                    self.seqInter.append(" ")
-        self.seqAEnd = self.seqAStart + len(self.seqA) - self.seqAGaps
-        self.seqBEnd = self.seqBStart + len(self.seqB) - self.seqBGaps
-
-    def getBottomSequence(self):
-        return self.seqB
-
-    def __repr__(self):
-        """
-        Object representation.
-        """
-        res = []
-        if self.isMultiple:
-            res.append("---------- Multi-Seq. Alignment ----------")
-        else:
-            res.append("---------- Alignment ----------")
-        res.append("Size       : " + str(self.size))
-        res.append("Type       : " + self.alignType)
-        res.append("Score      : " + str(self.alignScore))
-        if not self.isMultiple:
-            res.append("Identity   : " + str(self.identity) \
-                       + " ({0:.2f}%)".format(100 * (self.identity / len(self.seqA))))
-            res.append("Similarity : " + str(self.similarity) \
-                       + " ({0:.2f}%)".format(100 * (self.similarity / len(self.seqA))))
-        res.append("Gaps       : " + (str(self.gaps) if self.gaps > 0 else "None"))
-        res.append("")
-        if self.isMultiple:
-            res.append("PSSM : " + self.pssmDescription)
-            res.append("Aligned seq. : " + self.seqA.getDescription())
-            res.append("\t" + str(self.seqAGaps) + " Gaps, " + str(self.size - self.seqAGaps) + \
-                       " AAs (positions " + str(self.seqAStart) + " to " + str(self.seqAEnd) + ")")
-        else:
-            res.append("Upper seq. : " + self.seqA.getDescription())
-            res.append("\t" + str(self.seqAGaps) + " Gaps, " + str(self.size - self.seqAGaps) + \
-                       " AAs (positions " + str(self.seqAStart) + " to " + str(self.seqAEnd) + ")")
-            res.append("Lower seq. : " + self.seqB.getDescription())
-            res.append("\t" + str(self.seqBGaps) + " Gaps, " + str(self.size - self.seqBGaps) + \
-                       " AAs (positions " + str(self.seqBStart) + " to " + str(self.seqBEnd) + ")")
-        res.append("")
-
-        if self.condensed:
-            return "\n".join(res)
-
-        listA, listB, listI = [], [], []
-        maxALen, maxBLen = 0, 0
-        seqAPos, seqBPos = self.seqAStart, self.seqBStart
-        seqANextPos, seqBNextPos = seqAPos, seqBPos
-
-        for index in range(self.size):
-            a = self.seqA[index]
-            if not a.isGap():
-                seqANextPos += 1
-            listA.append(str(a))
-            maxALen = max(len(str(a)), maxALen)
-
-            if not self.isMultiple:
-                listI.append(self.seqInter[index])
-                b = self.seqB[index]
-                if not b.isGap():
-                    seqBNextPos += 1
-                listB.append(str(b))
-                maxBLen = max(len(str(b)), maxBLen)
-
-            if index != 0 and (index % self.chunkSize == 0 or index == len(self.seqA) - 1):
-                res.append(str(seqAPos))
-                for i in range(maxALen - 1, -1, -1):
-                    res.append("".join([s[-i - 1] if len(s) > i else " " for s in listA]))
-
-                if not self.isMultiple:
-                    res.append("".join(listI))
-                    res.append(str(seqBPos))
-                    for i in range(maxBLen):
-                        res.append("".join([s[i] if len(s) > i else " " for s in listB]))
-                res.append("\n")
-                listA, listB, listI = [], [], []
-                maxALen, maxBLen = 0, 0
-                seqAPos, seqBPos = seqANextPos, seqBNextPos
-
-        return "\n".join(res)
+from pyprot.base.sequence import Sequence
+from pyprot.align.aligned import Aligned
 
 
 class Align:
     """
-    Represents an alignment matrix, used to determine the alignment score between two sequences
+    Represents an alignment matrix, used to determine the alignment score between two Sequences
     """
 
     def __init__(self, scoreMatrix):
@@ -149,17 +15,17 @@ class Align:
         """
         self._scoreMatrix = scoreMatrix  # scoring matrix to use
 
-        self._colSeq = None  # copies of sequences to align
+        self._colSeq = None  # copies of Sequences to align
         self._rowSeq = None
 
-        self._alignMatrix = None  # Used for alignment scores between sequences
+        self._alignMatrix = None  # Used for alignment scores between Sequences
         self._rowGapMatrix = None  # used for affine gap penalty
         self._colGapMatrix = None
         self._originMatrix = None  # Used for backtracking purposes
 
         self._alignMode = ""  # Align mode (global, semiglobal, local)
         self._isSuboptimal = False  # Is the alignment subobtimal ?
-        self._isMultiple = False  # Is the alignment done against multiple sequences at once ?
+        self._isMultiple = False  # Is the alignment done against multiple Sequences at once ?
         self._resultCount = None  # The number of expected results
         self._subOptimalDepth = None
 
@@ -169,21 +35,21 @@ class Align:
         self._maxScoreRows = []  # Index of maximum values from _alignMatrix
         self._maxScoreCols = []
         self._currentAlignPath = []
-        self._bestAlignPath = []  # sequence of origins for alignment with best score
-        self._allAlignPaths = []  # all sequences of origins for alignments with best scores
+        self._bestAlignPath = []  # Sequence of origins for alignment with best score
+        self._allAlignPaths = []  # all Sequences of origins for alignments with best scores
 
         self._iniGapPenalty = 0  # Initial gap penalty
         self._extGapPenalty = 0  # Extended gap penalty
 
-        self._alignedRowSeq = None  # Aligned sequences (result of alignment)
+        self._alignedRowSeq = None  # Aligned Sequences (result of alignment)
         self._alignedColSeq = None
 
     def globalAlign(self, seqA, seqB, iniGapPenalty=1, extGapPenalty=None, semiGlobal=False, resultCount=-1):
         """
         Returns all best global alignment between 'seqA' and 'seqB' with the provided
         initial and extended gap penalties.
-        If 'semiGlobal' is True, allows alignment to discard the start and end of either sequence.
-        (this allows for better alignments when sequences overlap only partially)
+        If 'semiGlobal' is True, allows alignment to discard the start and end of either Sequence.
+        (this allows for better alignments when Sequences overlap only partially)
         """
         if semiGlobal:
             self._alignMode = "semiglobal"
@@ -239,15 +105,15 @@ class Align:
             for row, col in zip(self._maxScoreRows, self._maxScoreCols):
                 yield from self.__align(row, col)
 
-    def multiAlign(self, sequence, resultCount=1, subOptimalDepth=3):
+    def multiAlign(self, Sequence, resultCount=1, subOptimalDepth=3):
         """
-        Returns n best local M.S.A. of 'sequence' against the ScoreMatrix (which must be 
+        Returns n best local M.S.A. of 'Sequence' against the ScoreMatrix (which must be 
         position-specific and provide gap penalties) where n equals 'resultCount' (-1 for all).
         If 'subOptimalDepth' equals m, looks for m best suboptimal alignments as well.
         """
         self._isMultiple = True
         colSeq = [i for i in range(len(self._scoreMatrix))]
-        yield from self.localAlign(colSeq, sequence, 0, None, resultCount, subOptimalDepth)
+        yield from self.localAlign(colSeq, Sequence, 0, None, resultCount, subOptimalDepth)
 
     def __initialize(self, seqA, seqB, iniGapPenalty, extGapPenalty):
         """
@@ -265,7 +131,7 @@ class Align:
 
         self._alignedRowSeq = Sequence()
         if self._isMultiple:
-            # In MSA (align with PSSM) there is no column sequence
+            # In MSA (align with PSSM) there is no column Sequence
             self._alignedColSeq = []
         else:
             self._alignedColSeq = Sequence()
@@ -416,12 +282,12 @@ class Align:
 
             # Create result (Sequence obj. for MSA, Aligned obj. otherwise)
             if self._isMultiple:
-                result = Aligned(self._alignedColSeq, \
-                                 Sequence(self._alignedRowSeq, self._rowSeq.getDescription()), j, i, \
+                result = Aligned(self._alignedColSeq,
+                                 Sequence(self._alignedRowSeq, self._rowSeq.getDescription()), j, i,
                                  alignDescription, self._currentAlignScore, self._scoreMatrix, True)
             else:
-                result = Aligned(Sequence(self._alignedColSeq, self._colSeq.getDescription()), \
-                                 Sequence(self._alignedRowSeq, self._rowSeq.getDescription()), j, i, \
+                result = Aligned(Sequence(self._alignedColSeq, self._colSeq.getDescription()),
+                                 Sequence(self._alignedRowSeq, self._rowSeq.getDescription()), j, i,
                                  alignDescription, self._currentAlignScore, self._scoreMatrix, False)
 
             # Remember first best alignment for subobtimal lookup
@@ -469,7 +335,7 @@ class Align:
         sepSize = 5
         rep = {"": "", "L": "_   ", "D": " \\  ", "T": "   |", "LD": "_\\  ", "DT": " \\ |", "LT": "_  |",
                "LDT": "_\\ |"}
-        # Amino Acid column sequence
+        # Amino Acid column Sequence
         result.append(" " * 2 * sepSize)
         for aa in self._colSeq:
             result[-1] += '{a!s:<{w}}'.format(a=aa, w=sepSize)
@@ -483,7 +349,7 @@ class Align:
         for value in self._alignMatrix[0]:
             result[-1] += '{v:<{w}}'.format(v=value, w=sepSize)
 
-        # Amino Acid line sequence and values
+        # Amino Acid line Sequence and values
         for values, origins, aa in zip(self._alignMatrix[1:], self._originMatrix[1:], self._rowSeq):
             result.append(" ")
             for origin in origins:
