+++ pyprot/base/sequence.py
