+++ pyprot/base/test/test_sequence.py
